requirements.txt
praw
kafka-python
pyspark
textblob
psycopg2-binary
streamlit
matplotlib
pandas
